refactor(agent): extract type and error-class surface from dkg-agent.ts
Moves every type, interface, and error class previously declared in
`dkg-agent.ts` into a dedicated `dkg-agent-types.ts` module — including
the externally-visible public surface (`DKGAgentConfig`, `PublishOpts`,
`PublishAsyncOpts`, `PublishAsyncQuadEnvelope`, `PublishAsyncContent`,
`ContextGraphSub`, `ContextGraph{Subscription,Membership}{Record,Store}`,
`Peer{Health,ConnectionSnapshot,Diagnostics}`, `ChatSendResult`,
`{Durable,SharedMemory,Catchup}SyncDiagnostics`, `CclPublished*`,
`ContextGraphNotFoundError`, `InvalidContentError`) and the file-local
structural types the class consumes (`PreSignedAuthorAttestation`,
`LocalSwmSenderKey{Send,Receive}State`, `SyncRequestEnvelope`,
`RandomSamplingStartResult`, `ACKSignerResolution`, `SyncAccessDeniedError`,
`{Durable,SharedMemory}SyncResult`).

`dkg-agent.ts` keeps a re-export of the public symbols so external
imports of `./dkg-agent.js` continue to resolve. `packages/agent/src/index.ts`
now imports the public types from the new file directly.

Verified field-by-field against `HEAD~1` — every property, JSDoc, and
extends-clause is preserved; the only diff is the addition of the
`export` keyword where the original had file-local visibility.
